fix: sealed class LangString permits DirectionalLangString
